+++ app/src/main/java/com/trill/ecommerce/model/SalesAgentModel.kt
@@ -1,0 +1,8 @@
+package com.trill.ecommerce.model
+
+class SalesAgentModel() {
+    var uid: String? = null
+    var name: String? = null
+    var phone: String? = null
+    var image: String? = null
+}