+++ app/src/main/java/com/trill/ecommerce/notifications/TrillFirebaseMessagingService.kt
@@ -75,4 +75,5 @@ class TrillFirebaseMessagingService : FirebaseMessagingService() {
         notificationManager.notify(0, builder.build())
 
     }
+
 }