+++ app/src/main/java/com/trill/ecommerce/screens/checkout/mpesa/MpesaListener.kt
@@ -1,0 +1,7 @@
+package com.trill.ecommerce.screens.checkout.mpesa
+
+interface MpesaListener {
+    fun sendingSuccessful(transactionAmount: String, phoneNumber: String, transactionDate: String, MPesaReceiptNo: String)
+
+    fun sendingFailed(cause: String)
+}