+++ app/src/main/java/com/trill/ecommerce/util/HistoryItemClick.kt
@@ -1,0 +1,6 @@
+package com.trill.ecommerce.util
+
+import com.trill.ecommerce.model.OrderModel
+
+class HistoryItemClick(var isSuccess: Boolean, var category: OrderModel) {
+}