+++ app/src/main/java/com/trill/ecommerce/viewmodel/HistoryViewModel.kt
@@ -1,0 +1,4 @@
+package com.trill.ecommerce.viewmodel
+
+class HistoryViewModel {
+}